requirements.txt
python-docx
jsonschema
openai
python-dotenv
